Support for rescheduling appointment (#9956)
