+++ public/locale/en.json
@@ -396,9 +396,12 @@
   "app_settings": "App Settings",
   "apply": "Apply",
   "appointment_booking_success": "Your appointment has been successfully booked!",
+  "appointment_cancelled": "Appointment has been cancelled!",
   "appointment_created_success": "Appointment created successfully",
   "appointment_details": "Appointment Details",
   "appointment_not_found": "Appointment not found",
+  "appointment_rescheduled": "Appointment has been rescheduled!",
+  "appointment_rescheduled_successfully": "Appointment rescheduled successfully",
   "appointment_type": "Appointment Type",
   "appointments": "Appointments",
   "approve": "Approve",
@@ -1756,6 +1759,9 @@
   "required": "Required",
   "required_quantity": "Required Quantity",
   "reschedule": "Reschedule",
+  "reschedule_appointment": "Reschedule Appointment",
+  "rescheduled": "Rescheduled",
+  "rescheduling": "Rescheduling...",
   "resend_otp": "Resend OTP",
   "reset": "Reset",
   "reset_password": "Reset Password",
