+++ src/pages/Appointments/BookAppointment.tsx
@@ -1,5 +1,4 @@
 import { useMutation, useQuery } from "@tanstack/react-query";
-import { format, isBefore, isSameDay, startOfToday } from "date-fns";
 import { navigate } from "raviger";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
@@ -7,11 +6,8 @@ import { toast } from "sonner";
 
 import { cn } from "@/lib/utils";
 
-import Calendar from "@/CAREUI/interactive/Calendar";
-
 import { Button } from "@/components/ui/button";
 import { Label } from "@/components/ui/label";
-import { ScrollArea } from "@/components/ui/scroll-area";
 import {
   Select,
   SelectContent,
@@ -19,7 +15,6 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import { Separator } from "@/components/ui/separator";
 import { Textarea } from "@/components/ui/textarea";
 
 import { Avatar } from "@/components/Common/Avatar";
@@ -29,12 +24,10 @@ import useAppHistory from "@/hooks/useAppHistory";
 
 import mutate from "@/Utils/request/mutate";
 import query from "@/Utils/request/query";
-import { dateQueryString, formatDisplayName, formatName } from "@/Utils/utils";
-import {
-  groupSlotsByAvailability,
-  useAvailabilityHeatmap,
-} from "@/pages/Appointments/utils";
+import { formatDisplayName, formatName } from "@/Utils/utils";
 import scheduleApis from "@/types/scheduling/scheduleApis";
+
+import { AppointmentSlotPicker } from "./components/AppointmentSlotPicker";
 
 interface Props {
   facilityId: string;
@@ -46,11 +39,9 @@ export default function BookAppointment(props: Props) {
   const { goBack } = useAppHistory();
 
   const [resourceId, setResourceId] = useState<string>();
-  const [selectedMonth, setSelectedMonth] = useState(new Date());
-  const [selectedDate, setSelectedDate] = useState(new Date());
+  const [selectedSlotId, setSelectedSlotId] = useState<string>();
 
   const [reason, setReason] = useState("");
-  const [selectedSlotId, setSelectedSlotId] = useState<string>();
 
   const resourcesQuery = useQuery({
     queryKey: ["availableResources", props.facilityId],
@@ -62,31 +53,6 @@ export default function BookAppointment(props: Props) {
   });
   const resource = resourcesQuery.data?.users.find((r) => r.id === resourceId);
 
-  const heatmapQuery = useAvailabilityHeatmap({
-    facilityId: props.facilityId,
-    userId: resourceId,
-    month: selectedMonth,
-  });
-
-  const slotsQuery = useQuery({
-    queryKey: [
-      "slots",
-      props.facilityId,
-      resourceId,
-      dateQueryString(selectedDate),
-    ],
-    queryFn: query(scheduleApis.slots.getSlotsForDay, {
-      pathParams: { facility_id: props.facilityId },
-      body: {
-        // voluntarily coalesce to empty string since we know query would be
-        // enabled only if resourceId and selectedDate are present
-        user: resourceId ?? "",
-        day: dateQueryString(selectedDate),
-      },
-    }),
-    enabled: !!resourceId && !!selectedDate,
-  });
-
   const { mutateAsync: createAppointment } = useMutation({
     mutationFn: mutate(scheduleApis.slots.createAppointment, {
       pathParams: {
@@ -95,89 +61,6 @@ export default function BookAppointment(props: Props) {
       },
     }),
   });
-
-  const renderDay = (date: Date) => {
-    const isSelected = isSameDay(date, selectedDate);
-    const isBeforeToday = isBefore(date, startOfToday());
-
-    const availability = heatmapQuery.data?.[dateQueryString(date)];
-
-    if (
-      heatmapQuery.isFetching ||
-      !availability ||
-      availability.total_slots === 0 || // TODO: replace this with stripes. -- indicates day is not available
-      isBeforeToday
-    ) {
-      return (
-        <button
-          disabled
-          onClick={() => {
-            setSelectedDate(date);
-            setSelectedSlotId(undefined);
-          }}
-          className={cn(
-            "h-full w-full hover:bg-gray-50 rounded-lg relative overflow-hidden border border-gray-200",
-            isSelected ? "ring-2 ring-primary-500" : "",
-          )}
-        >
-          <div className="relative z-10">
-            <span>{date.getDate()}</span>
-          </div>
-        </button>
-      );
-    }
-
-    const { booked_slots, total_slots } = availability;
-    const bookedPercentage = booked_slots / total_slots;
-    const tokensLeft = total_slots - booked_slots;
-    const isFullyBooked = tokensLeft <= 0;
-
-    return (
-      <button
-        disabled={isBeforeToday || isFullyBooked}
-        onClick={() => {
-          setSelectedDate(date);
-          setSelectedSlotId(undefined);
-        }}
-        className={cn(
-          "h-full w-full hover:bg-gray-50 rounded-lg relative overflow-hidden border-2 hover:scale-105 hover:shadow-md transition-all",
-          isSelected ? "border-primary-500" : "border-gray-200",
-          isFullyBooked ? "bg-gray-200" : "bg-white",
-        )}
-      >
-        <div className="relative z-10">
-          <span>{date.getDate()}</span>
-          {Number.isFinite(tokensLeft) && (
-            <span
-              className={cn(
-                "text-xs text-gray-500 block font-semibold",
-                bookedPercentage >= 0.8
-                  ? "text-red-500"
-                  : bookedPercentage >= 0.5
-                    ? "text-yellow-500"
-                    : "text-primary-500",
-              )}
-            >
-              {tokensLeft} left
-            </span>
-          )}
-        </div>
-        {!isFullyBooked && (
-          <div
-            className={cn(
-              "absolute bottom-0 left-0 w-full transition-all",
-              bookedPercentage > 0.8
-                ? "bg-red-100"
-                : bookedPercentage > 0.5
-                  ? "bg-yellow-100"
-                  : "bg-primary-100",
-            )}
-            style={{ height: `${Math.min(bookedPercentage * 100, 100)}%` }}
-          />
-        )}
-      </button>
-    );
-  };
 
   const handleSubmit = async () => {
     if (!resourceId) {
@@ -268,103 +151,12 @@ export default function BookAppointment(props: Props) {
               !resourceId && "opacity-50 pointer-events-none",
             )}
           >
-            <div>
-              <Calendar
-                month={selectedMonth}
-                onMonthChange={(month) => {
-                  setSelectedMonth(month);
-                  setSelectedSlotId(undefined);
-                }}
-                renderDay={renderDay}
-                className="mb-6"
-                highlightToday={false}
-              />
-            </div>
-
-            <div>
-              <div className="flex items-center justify-between mb-4">
-                <h3 className="font-medium">{t("available_time_slots")}</h3>
-              </div>
-              <ScrollArea>
-                <div className="max-h-96">
-                  {slotsQuery.data == null && (
-                    <div className="flex items-center justify-center py-32 border-2 border-gray-200 border-dashed rounded-lg text-center">
-                      <p className="text-gray-400">
-                        {t("to_view_available_slots_select_resource_and_date")}
-                      </p>
-                    </div>
-                  )}
-                  {slotsQuery.data?.results.length === 0 && (
-                    <div className="flex items-center justify-center py-32 border-2 border-gray-200 border-dashed rounded-lg text-center">
-                      <p className="text-gray-400">
-                        {t("no_slots_available_for_this_date")}
-                      </p>
-                    </div>
-                  )}
-                  {!!slotsQuery.data?.results.length &&
-                    groupSlotsByAvailability(slotsQuery.data.results).map(
-                      ({ availability, slots }) => (
-                        <div key={availability.name}>
-                          <h4 className="text-lg font-semibold mb-3">
-                            {availability.name}
-                          </h4>
-                          <div className="flex flex-wrap gap-2">
-                            {slots.map((slot) => {
-                              const percentage =
-                                slot.allocated / availability.tokens_per_slot;
-
-                              return (
-                                <Button
-                                  key={slot.id}
-                                  size="lg"
-                                  variant={
-                                    selectedSlotId === slot.id
-                                      ? "outline_primary"
-                                      : "outline"
-                                  }
-                                  onClick={() => {
-                                    if (selectedSlotId === slot.id) {
-                                      setSelectedSlotId(undefined);
-                                    } else {
-                                      setSelectedSlotId(slot.id);
-                                    }
-                                  }}
-                                  disabled={
-                                    slot.allocated ===
-                                    availability.tokens_per_slot
-                                  }
-                                  className="flex flex-col items-center group gap-0"
-                                >
-                                  <span className="font-semibold">
-                                    {format(slot.start_datetime, "HH:mm")}
-                                  </span>
-                                  <span
-                                    className={cn(
-                                      "text-xs group-hover:text-inherit",
-                                      percentage >= 1
-                                        ? "text-gray-400"
-                                        : percentage >= 0.8
-                                          ? "text-red-600"
-                                          : percentage >= 0.6
-                                            ? "text-yellow-600"
-                                            : "text-green-600",
-                                    )}
-                                  >
-                                    {availability.tokens_per_slot -
-                                      slot.allocated}{" "}
-                                    left
-                                  </span>
-                                </Button>
-                              );
-                            })}
-                          </div>
-                          <Separator className="my-6" />
-                        </div>
-                      ),
-                    )}
-                </div>
-              </ScrollArea>
-            </div>
+            <AppointmentSlotPicker
+              facilityId={props.facilityId}
+              resourceId={resourceId}
+              selectedSlotId={selectedSlotId}
+              onSlotSelect={setSelectedSlotId}
+            />
           </div>
 
           <div className="flex justify-end gap-4">
